requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cssselect==1.0.3
fake-useragent==0.1.11
future==0.17.1
html5lib==1.0.1
idna==2.8
lxml==4.3.4
parse==1.12.0
pycparser==2.19
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
python-telegram-bot==11.1.0
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.1
tqdm==4.32.1
urllib3==1.25.3
w3lib==1.20.0
webencodings==0.5.1
websockets==7.0
